Saner whitespace rules for AST templates
The whitespace rules for AST matching were long a source of confusion.
This tries to do better, by conceptually computing the diff between the
whitespace in the template and that in the replacement and attempting
to apply that to the match. Fixes #30.

While we're at it, also canonicalize introduced keyword arguments to not
have whitespace around the `=`.